:fire: Reorganize test namespaces.
