+++ tests/Helper/ExporterPlugin.php
@@ -17,7 +17,7 @@
  * @license   http://www.opensource.org/licenses/mit-license.php The MIT License (MIT)
  * */
 
-namespace NelsonMartell\Test\plugins;
+namespace NelsonMartell\Test\Helpers;
 
 use NelsonMartell as NML;
 use NelsonMartell\Extensions\String;
