+++ tests/Helper/IComparableTestHelper.php
@@ -17,14 +17,14 @@
  * @license   http://www.opensource.org/licenses/mit-license.php The MIT License (MIT)
  * */
 
-namespace NelsonMartell\Test;
+namespace NelsonMartell\Test\Helpers;
 
 use NelsonMartell as NML;
 use NelsonMartell\Extensions\String;
 use NelsonMartell\IComparable;
 use NelsonMartell\Object;
 use NelsonMartell\Version;
-use NelsonMartell\Test\plugins\ExporterPlugin;
+use NelsonMartell\Test\Helpers\ExporterPlugin;
 use \InvalidArgumentException;
 
 /**
