+++ tests/TestCase/ObjectTest.php
@@ -17,13 +17,13 @@
  * @license   http://www.opensource.org/licenses/mit-license.php The MIT License (MIT)
  * */
 
-namespace NelsonMartell\Test;
+namespace NelsonMartell\Test\TestCase;
 
 use NelsonMartell as NML;
 use NelsonMartell\Extensions\String;
 use NelsonMartell\Object;
 use NelsonMartell\Version;
-use NelsonMartell\Test\plugins\ExporterPlugin;
+use NelsonMartell\Test\Helpers\ExporterPlugin;
 use \PHPUnit_Framework_TestCase as TestCase;
 use \InvalidArgumentException;
